First pass at migrating to Nokee plugins
We didn't try the build on Windows yet to validate the Visual Studio IDE
experience. Building should work as expected.

The legacy installer was highly approximated from installUtils.gradle.
Instead of staging the binaries from the producer projects, we expose the
installer elements and consume them in the installer project.

One of the rough edges is the ability to create custom artifacts and
configuration based on a variant. The work is planned but not started.
This is a considerable work that can be worked around for the time
being. Depending on the end result, the workaround may differ.

We can add the base name property on the extensions to bridge a gap with
Gradle core support in a nightly.

Mixed language requires to use two distinct project at the moment. This
is not how we want to solve the use case. There is ways to add more
language to a component but it's extremely verbose and use deep internal
of the Nokee plugins. It would be best to introduce a generic, language
agnostic components. Everything is in place, the bottleneck is the test
coverage of such scenario.