Introduce file tree isolation
For C/C++ project that trigger the fallback behavior for walking the
headers AND one header search path is at the project dir, Gradle would
try to snapshot all files in the header search paths which would include
the build directory with all sorts of files that will change with each
build. This will prevent the task from ever being up-to-date. The
isolation will sync the header to another directory (isolated) before
adding that directory as the header search path.